Update 2020-02-14-golang-roguelike.md
Shouldn't grid be a considered an Entity? (Drawable, Renderable)
I'm also working on a minimal ECS and your thoughts on it is helping me with the architecture.
But classic ECS does not have data or methods attached to Entity. Entity is merely an unique ID.

If you're interested in my ecs, perhaps we can develop it together.